requirements.txt
appdirs==1.4.4
arrow==0.17.0
asgiref==3.2.7
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==3.0.6
django-crispy-forms==1.9.1
django-cuser==2017.3.16
django-extensions==3.1.0
django-filter==2.3.0
django-jalali-date==0.3.1
django-phonenumber-field==5.0.0
django-widget-tweaks==1.4.8
idna==2.10
isodate==0.6.0
jdatetime==3.6.2
lxml==4.6.2
phonenumbers==8.12.16
Pillow==7.1.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
zeep==4.0.0
